Replace Pack plugin info with link
It wasn't clear (to me anyway) that the Packs needed to be downloaded and installed before I could associate a device with a project. The Pack plug-in installation page is very clear though so I have removed Pack plugin information from this page and linked to that one. Hope that's OK.